Organize files & class names
