+++ Sources/BraintreePayPal/BillingAgreementMetadata/BTPayPalBillingAgreementPlanType.swift
@@ -1,0 +1,8 @@
+// AKA charge pattern
+// TODO docs
+public enum BTPayPalBillingAgreementPlanType: String {
+    case recurring = "RECURRING"
+    case installment = "INSTALLMENT"
+    case unscheduled = "UNSCHEDULED"
+    case subscription = "SUBSCRIPTION"
+}
